create component mobile organism ReviewSkills and update component template Skill
