--- packages/@coorpacademy-components/src/template/app-review/skills/prop-types.ts
@@ -1,7 +1,0 @@
-import {SkillCardProps} from '../../../molecule/skill-card/prop-types';
-
-export type SkillProps = Omit<SkillCardProps, 'aria-label' | 'buttonLabel' | 'isCustom'>;
-
-export type ItemProps = {
-  item: SkillProps;
-};
