--- packages/@coorpacademy-components/src/template/app-review/skills/stories.tsx
@@ -1,16 +1,0 @@
-// import * as React from 'react';
-// import {storiesOf} from '@storybook/react-native';
-// import RevisionHome from './revision-home';
-
-// const emptySkills: [] = [];
-// const skillsToRevise: {title: string; info: string}[] = [
-//   {title: 'Negotiation and powers of persuasion', info: '75 questions to revise'},
-//   {title: 'Negotiation and powers of persuasion', info: '75 questions to revise'},
-//   {title: 'Negotiation and powers of persuasion', info: '75 questions to revise'},
-//   {title: 'Negotiation and powers of persuasion', info: '75 questions to revise'},
-//   {title: 'Creativity and innovation', info: '75 questions to revise'}
-// ];
-
-// storiesOf('RevisionHome', module)
-//   .add('No skills', () => <RevisionHome skillsToRevise={emptySkills} />)
-//   .add('Few skills', () => <RevisionHome skillsToRevise={skillsToRevise} />);
